requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==23.1.0
cachetools==5.5.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fastapi==0.115.12
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-vision==3.7.4
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
h5py==3.12.1
idna==3.6
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.0.0
pillow==10.4.0
proto-plus==1.24.0
protobuf>=5.26.1,<6.0.0 
# dependencies
firebase-admin
httpxpip install firebase-admin httpx
pyasn1>=0.4.1,<0.5.0
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.18.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.32.3
rich==13.9.2
rsa==4.9
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.46.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow==2.18.0
termcolor==2.4.0
trio==0.23.1
trio-websocket==0.11.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.34.0
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
pydantic-settings